Start working on class Snake
