+++ SnakeMess/Snake.csproj
@@ -46,6 +46,7 @@
     <Compile Include="Factory.cs" />
     <Compile Include="Position.cs" />
     <Compile Include="Snake.cs" />
+    <Compile Include="SnakeGame.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
